requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
beautifulsoup4==4.13.4
certifi==2025.7.14
charset-normalizer==3.4.2
colorama==0.4.6
crispy-bootstrap5==2025.6
distro==1.9.0
Django==5.2.4
django-ckeditor==6.7.3
django-crispy-forms==2.4
django-js-asset==3.1.2
django-widget-tweaks==1.5.0
faiss-cpu==1.11.0
filelock==3.18.0
fsspec==2025.5.1
groq==0.30.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.4
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.1
packaging==25.0
pillow==11.3.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
sentence-transformers==5.0.0
setuptools==80.9.0
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.5.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.2
torch==2.7.1
tqdm==4.67.1
transformers==4.53.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
